Deleted duplicates

+++ core/src/main/kotlin/com/stepanov/bbf/bugfinder/reducer/ResultsFilter.kt
@@ -24,21 +24,25 @@ object ResultsFilter {
                     return@forEach
                 }
 
-                if (file.readText().contains("java.lang.OutOfMemoryError: Cannot reserve")) {
+                val text = file.readText()
+
+                if (text.contains("java.lang.OutOfMemoryError: Cannot reserve")) {
                     file.delete()
                     return@forEach
                 }
 
-                val stackTrace = extractStackTrace(file)
-                if (stackTrace.count { it == '\n' } > 5) { // in empty stack trace not really interested right now
-                    File(CompilerArgs.resultsDir).walkTopDown().filter { it.isFile }.forEach { other ->
-                        if (other.name != file.name) {
-                            val st2 = extractStackTrace(other)
-                            if (
-                                deleteSourceFile(stackTrace) == deleteSourceFile(st2)
-                            ) {
-                                println("${other.name} -> ${file.name}")
-                                other.delete()
+                if (text.contains("COMPILER_CRASHED")) {
+                    val stackTrace = extractStackTrace(file)
+                    if (stackTrace.count { it == '\n' } > 5) { // in empty stack trace not really interested right now
+                        File(CompilerArgs.resultsDir).walkTopDown().filter { it.isFile }.forEach { other ->
+                            if (other.name != file.name) {
+                                val st2 = extractStackTrace(other)
+                                if (
+                                    deleteSourceFile(stackTrace) == deleteSourceFile(st2)
+                                ) {
+                                    println("${other.name} -> ${file.name}")
+                                    other.delete()
+                                }
                             }
                         }
                     }
@@ -51,8 +55,8 @@ object ResultsFilter {
         (if (file.name.contains("FRONTEND"))
             file.readText().substringAfterLast("causeThrowable")
         else
-            file.readText().substringAfterLast("STACKTRACE"))
-            .substringAfter("// \tat")
+            file.readText().substringAfterLast("===================="))
+            .substringAfter("\tat")
             .split("\n").take(10).joinToString("\n")
 
     private fun deleteSourceFile(code: String) = code.substringBefore("com.stepanov.bbf")
